Move existing functions to centralized history file
